feat: add Pull Request Template
Added a new Pull Request template with sections for specifying the type
of Pul Request (Refactor, Feature, Bug Fix, Optimization, Documentation
Update, CI/CD), a concise description of changes, related tickets, QA
instructions with screenshots/recordings, resources for reviewers, and a
checkbox for tests with a target of 90% code coverage. Also included a
field to specify the target branch for merging.