fix: separate event setting from button creation
Fixes the bug of not setting events on a button because the button is
not yet created or appended to the DOM. Event setting relies on the
element's ID in the DOM. Separated event setting into another method to
address this issue. Additionally, return `this` in methods where
chaining is necessary.