feat: implement container component (#13)
* feat: implement Container component

Introduce the Container component to handle the creation and removal of
container elements. This component ensures that the provided config is
verified using the `isConfigVerified` utility. The `create` method sets
the attributes for the container, while the `remove` method handles the
removal of the container from the DOM.

* test: add tests for Container component

Adds tests for the Container component using JSDOM and Jest. Tests
include container creation, text or element appending, and container
removal. Ensures proper functionality of Container's create and remove
methods.